Add rate to user profile impl and design

+++ app/src/main/java/com/krygodev/appforartists/core/domain/model/ImageModel.kt
@@ -12,5 +12,5 @@ data class ImageModel(
     val tags: List<String> = listOf(),
     val starredBy: List<String> = listOf(),
     val starsSum: Int = 0,
-    val starsAvg: Int = 0
+    val starsAvg: Float = 0f
 )
